# backend/api.yaml
openapi: 3.0.0
paths:
...
  /api/v1/categories:
    post:
      description: Tạo một danh mục sản phẩm mới trong hệ thống với tên và mô tả
      operationId: CategoryController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryDto'
      responses:
        '201':
          description: Danh mục được tạo thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  category_id:
                    type: number
                    description: ID của danh mục
                  name:
                    type: string
                    description: Tên danh mục
                  description:
                    type: string
                    description: Mô tả danh mục
                    nullable: true
                  is_active:
                    type: boolean
                    description: Trạng thái hoạt động
                  created_at:
                    type: string
                    format: date-time
                    description: Thời gian tạo
                  updated_at:
                    type: string
                    format: date-time
                    description: Thời gian cập nhật
        '400':
          description: Yêu cầu không hợp lệ - Dữ liệu đầu vào không đúng định dạng
        '401':
          description: Chưa xác thực - Token không hợp lệ
        '403':
          description: >-
            Không có quyền truy cập - Chỉ MANAGER và STAFF mới có thể tạo danh
            mục
        '409':
          description: Xung đột dữ liệu - Tên danh mục đã tồn tại
      security:
        - JWT-auth: []
      summary: Tạo danh mục mới - Chỉ MANAGER và STAFF
      tags:
        - categories
    get:
      description: Trả về danh sách tất cả danh mục sản phẩm với hỗ trợ phân trang
      operationId: CategoryController_findAll_v1
      parameters:
        - name: page
          required: false
          in: query
          description: 'Số trang (mặc định: 1)'
          schema:
            minimum: 1
            default: 1
            example: 1
            type: number
        - name: limit
          required: false
          in: query
          description: 'Số lượng bản ghi trên mỗi trang (mặc định: 10)'
          schema:
            minimum: 1
            maximum: 100
            default: 10
            example: 10
            type: number
      responses:
        '200':
          description: Danh sách danh mục được phân trang thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        category_id:
                          type: number
                        name:
                          type: string
                        description:
                          type: string
                          nullable: true
                        is_active:
                          type: boolean
                        created_at:
                          type: string
                          format: date-time
                        updated_at:
                          type: string
                          format: date-time
                  pagination:
                    $ref: '#/components/schemas/PaginationMetadata'
        '401':
          description: Chưa xác thực - Token không hợp lệ
      security:
        - JWT-auth: []
      summary: Lấy danh sách danh mục với phân trang - Tất cả role
      tags:
        - categories
  /api/v1/categories/{id}:
    get:
      description: Lấy thông tin chi tiết của một danh mục cụ thể bằng ID
      operationId: CategoryController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          description: ID của danh mục cần lấy thông tin
          schema:
            example: 1
            type: number
      responses:
        '200':
          description: Thông tin chi tiết danh mục
          content:
            application/json:
              schema:
                type: object
                properties:
                  category_id:
                    type: number
                    description: ID của danh mục
                  name:
                    type: string
                    description: Tên danh mục
                  description:
                    type: string
                    description: Mô tả danh mục
                    nullable: true
                  is_active:
                    type: boolean
                    description: Trạng thái hoạt động
                  created_at:
                    type: string
                    format: date-time
                    description: Thời gian tạo
                  updated_at:
                    type: string
                    format: date-time
                    description: Thời gian cập nhật
        '401':
          description: Chưa xác thực - Token không hợp lệ
        '404':
          description: Không tìm thấy danh mục với ID được cung cấp
      security:
        - JWT-auth: []
      summary: Lấy thông tin danh mục theo ID - Tất cả role
      tags:
        - categories
    patch:
      description: >-
        Cập nhật thông tin của một danh mục cụ thể như tên, mô tả hoặc trạng
        thái hoạt động
      operationId: CategoryController_update_v1
      parameters:
        - name: id
          required: true
          in: path
          description: ID của danh mục cần cập nhật
          schema:
            example: 1
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryDto'
      responses:
        '200':
          description: Danh mục được cập nhật thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  category_id:
                    type: number
                    description: ID của danh mục
                  name:
                    type: string
                    description: Tên danh mục đã cập nhật
                  description:
                    type: string
                    description: Mô tả danh mục đã cập nhật
                    nullable: true
                  is_active:
                    type: boolean
                    description: Trạng thái hoạt động đã cập nhật
                  updated_at:
                    type: string
                    format: date-time
                    description: Thời gian cập nhật
        '400':
          description: Yêu cầu không hợp lệ - Dữ liệu đầu vào không đúng định dạng
        '401':
          description: Chưa xác thực - Token không hợp lệ
        '403':
          description: >-
            Không có quyền truy cập - Chỉ MANAGER và STAFF mới có thể cập nhật
            danh mục
        '404':
          description: Không tìm thấy danh mục với ID được cung cấp
        '409':
          description: Xung đột dữ liệu - Tên danh mục đã tồn tại
      security:
        - JWT-auth: []
      summary: Cập nhật thông tin danh mục - Chỉ MANAGER và STAFF
      tags:
        - categories
    delete:
      description: >-
        Xóa vĩnh viễn một danh mục khỏi hệ thống. Lưu ý: Không thể xóa danh mục
        đang có sản phẩm liên quan
      operationId: CategoryController_remove_v1
      parameters:
        - name: id
          required: true
          in: path
          description: ID của danh mục cần xóa
          schema:
            example: 1
            type: number
      responses:
        '200':
          description: Danh mục được xóa thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  category_id:
                    type: number
                    description: ID của danh mục đã xóa
                  name:
                    type: string
                    description: Tên danh mục đã xóa
                  message:
                    type: string
                    description: Thông báo xác nhận xóa thành công
        '401':
          description: Chưa xác thực - Token không hợp lệ
        '403':
          description: Không có quyền truy cập - Chỉ MANAGER mới có thể xóa danh mục
        '404':
          description: Không tìm thấy danh mục với ID được cung cấp
        '409':
          description: Xung đột dữ liệu - Danh mục đang được sử dụng bởi các sản phẩm khác
      security:
        - JWT-auth: []
      summary: Xóa danh mục theo ID - Chỉ MANAGER
      tags:
        - categories
  /api/v1/categories/bulk:
    delete:
      description: >-
        Xóa nhiều danh mục cùng lúc. Các danh mục có sản phẩm liên quan sẽ không
        thể xóa và sẽ được báo lỗi
      operationId: CategoryController_bulkDelete_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkDeleteCategoryDto'
      responses:
        '200':
          description: Quá trình xóa hàng loạt hoàn thành với kết quả chi tiết
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: array
                    items:
                      type: number
                    description: Danh sách ID các danh mục đã xóa thành công
                  failed:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: number
                          description: ID danh mục không thể xóa
                        reason:
                          type: string
                          description: Lý do không thể xóa
                    description: Danh sách các danh mục không thể xóa và lý do
                  summary:
                    type: object
                    properties:
                      total:
                        type: number
                        description: Tổng số danh mục được yêu cầu xóa
                      success:
                        type: number
                        description: Số danh mục xóa thành công
                      failed:
                        type: number
                        description: Số danh mục không thể xóa
        '400':
          description: Yêu cầu không hợp lệ - Danh sách ID không đúng định dạng
        '401':
          description: Chưa xác thực - Token không hợp lệ
        '403':
          description: Không có quyền truy cập - Chỉ MANAGER mới có thể xóa danh mục
      security:
        - JWT-auth: []
      summary: Xóa nhiều danh mục cùng lúc - Chỉ MANAGER
      tags:
        - categories
  /api/v1/categories/admin/test:
    get:
      description: >-
        Endpoint để kiểm tra xem Category Controller có hoạt động bình thường
        không (smoke test)
      operationId: CategoryController_adminTest_v1
      parameters: []
      responses:
        '200':
          description: Kiểm tra thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Thông báo xác nhận controller hoạt động bình thường
                    example: Category controller is working!
        '401':
          description: Chưa xác thực - Token không hợp lệ
        '403':
          description: Không có quyền truy cập - Chỉ MANAGER mới có thể thực hiện kiểm tra
      security:
        - JWT-auth: []
      summary: Kiểm tra hoạt động của Category Controller - Chỉ MANAGER
      tags:
        - categories
...
  /api/v1/products/category/{categoryId}:
    get:
      operationId: ProductController_findByCategory_v1
      parameters:
        - name: categoryId
          required: true
          in: path
          description: Category ID
          schema:
            type: number
        - name: page
          required: false
          in: query
          description: Số trang (bắt đầu từ 1)
          schema:
            minimum: 1
            default: 1
            example: 1
            type: number
        - name: limit
          required: false
          in: query
          description: Số lượng bản ghi trên mỗi trang
          schema:
            minimum: 1
            maximum: 100
            default: 10
            example: 10
            type: number
      responses:
        '200':
          description: Danh sách sản phẩm theo danh mục
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: '#/components/schemas/PaginationMetadata'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Insufficient permissions
      security:
        - JWT-auth: []
      summary: Lấy sản phẩm theo danh mục với pagination - TẤT CẢ ROLES
      tags:
        - products
...
info:
  title: Cake POS API
  description: API documentation for Cake POS System - Hệ thống quản lý cửa hàng bánh ngọt
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Enter JWT token
      in: header
  schemas:
...
    PaginationMetadata:
      type: object
      properties:
        page:
          type: number
          description: Số trang hiện tại
          example: 1
          minimum: 1
        limit:
          type: number
          description: Số lượng bản ghi trên mỗi trang
          example: 10
          minimum: 1
          maximum: 100
        total:
          type: number
          description: Tổng số bản ghi
          example: 50
          minimum: 0
        totalPages:
          type: number
          description: Tổng số trang
          example: 5
          minimum: 0
        hasNext:
          type: boolean
          description: Có trang tiếp theo hay không
          example: true
        hasPrev:
          type: boolean
          description: Có trang trước hay không
          example: false
      required:
        - page
        - limit
        - total
        - totalPages
        - hasNext
        - hasPrev
...
    CreateCategoryDto:
      type: object
      properties:
        name:
          type: string
          description: Tên danh mục
          example: Bánh kem
          maxLength: 100
        description:
          type: string
          description: Mô tả danh mục
          example: Các loại bánh kem tươi và ngon
          maxLength: 1000
      required:
        - name
    UpdateCategoryDto:
      type: object
      properties:
        name:
          type: string
          description: Tên danh mục
          example: Bánh kem
          maxLength: 100
        description:
          type: string
          description: Mô tả danh mục
          example: Các loại bánh kem tươi và ngon
          maxLength: 1000
    BulkDeleteCategoryDto:
      type: object
      properties:
        ids:
          description: Danh sách ID các danh mục cần xóa
          example:
            - 1
            - 2
            - 3
          type: array
          items:
            type: number
      required:
        - ids
...